Refactor `LogLikelihood`
