--- src/stats/log_likelihood/mod.rs
@@ -1,5 +1,0 @@
-mod discrete;
-mod gaussian;
-
-mod structs;
-pub use structs::*;
